Enable using blockscout for our testnet and localnet (#2)
Fixes: https://github.com/hokunet/hoku-deploy/issues/184

This is a refactor of the docker compose functionality of the upstream
blockscout to work with our testnet, as well as work against localnet
for development. I've include previous work from @sander and
@islamaliev.


Some notes:
- The state of this PR is currently running on the Piknik 0 node at
https://explorer.testnet.hoku.sh/
- The indexer has a hard time keeping up with our block rate. It seems
to catchup in spurts, I'm not sure if there is a way to configure it
differently to work better. Future work might include looking into this
and optimizing.
- There are _many_ different ways to [configure
blockscout](https://docs.blockscout.com/setup/env-variables). Future
work could include optimizations, as well as look and feel updates to
match our branding.

---------

Signed-off-by: Sander Pick <sanderpick@gmail.com>
Co-authored-by: Sander Pick <sanderpick@gmail.com>
Co-authored-by: Islam Aliev <aliev.islam@gmail.com>