feat: enable indexing internal and pending txns (#8)
The default IPC instructions for deploying a blockscout explorer for a
subnet suggested that internal and pending transactions should not be
indexed. For our subnet, having these kinds of transactions indexed
makes sense. Specifically, we want users of the explorer to be able to
see the values associated with token transfers, and more generally, any
internal transactions.

This PR updates the configuration so internal and pending transactions
are indexed.